requirements.txt
asgiref==3.2.7
Django==3.0
django-cors-headers==3.3.0
djangorestframework==3.11.0
pytz==2022.7.1
sqlparse==0.3.1
tzdata==2022.7
